구독 회차 보증금 월렌탈료 계산 로직 수정 및 결제 준비 단계에서 1회차 결제시 상품가격의 20% 보증금 + 월렌탈료 1회차 결제로 결제할 수 있게 수정

+++ src/main/java/com/imchobo/sayren_back/domain/order/dto/OrderItemResponseDTO.java
@@ -12,6 +12,6 @@ public class OrderItemResponseDTO {
   private Long orderItemId;       // 주문 아이템 PK
   private Long productId;         // 상품 ID
   private String productName;     // 상품명
-  private Integer priceSnapshot;  // 주문 시점 가격
+  private Long priceSnapshot;  // 주문 시점 가격
   private Long planId;            // 요금제 ID (null 가능)
 }
