+++ src/main/java/com/imchobo/sayren_back/domain/order/entity/OrderItem.java
@@ -31,6 +31,6 @@ public class OrderItem {
   private OrderPlan orderPlan;
 
   @Column(nullable = false)
-  private Integer productPriceSnapshot; // 주문 시점 상품 가격 (NOT NULL)
+  private Long productPriceSnapshot; // 주문 시점 상품 가격 (NOT NULL)
 //   주문 시점 상품 가격 (스냅샷)
 }
